requirements.txt
django
djangorestframework
django-apscheduler
python-dotenv
